+++ hedera-mirror-rosetta/app/domain/types/constants_test.go
@@ -31,7 +31,7 @@ import (
 func TestResponseCodeUpToDate(t *testing.T) {
 	// given
 	for code, name := range services.ResponseCodeEnum_name {
-		assert.Equal(t, name, TransactionResults[code])
+		assert.Equal(t, name, TransactionResults[int16(code)])
 	}
 }
